config: adding influx sections and meter variables
moving from commandline options to influx section to supply
influx clients with data.
adding read and connected to the device to have a default
list of measurements, and being able to tag or save some
information to what device was connected at the time.
this is important when used in a lab environment constantly
changing devices.